refactor: OpenAI 공식 SDK 마이그레이션 및 모델 업그레이드

+++ src/main/java/com/techrag/tech_rag_assitant/search/SearchService.java
@@ -32,7 +32,7 @@ public class SearchService {
         log.info("Searching for: {}", query);
 
         // 1. 질문 임베딩 생성
-        List<Double> queryEmbedding = embeddingService.createEmbedding(query);
+        List<Float> queryEmbedding = embeddingService.createEmbedding(query);
         String embeddingStr = embeddingService.embeddingToString(queryEmbedding);
 
         // 2. 유사도 검색
